requirements.txt
absl-py==1.3.0
addict==2.4.0
albumentations==1.3.0
alembic==1.10.4
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
bottle==0.12.25
cachetools==5.2.0
certifi
cffi==1.15.1
charset-normalizer==2.1.1
chex==0.1.5
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dm-tree==0.1.8
docker-pycreds==0.4.0
e==1.4.5
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.12.0
flatbuffers==2.0.7
flax==0.5.3
fonttools==4.38.0
fqdn==1.5.1
fsspec==2023.5.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
gpustat==1.0.0
greenlet==2.0.2
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.14.1
icecream==2.1.3
idna==3.4
imageio==2.23.0
importlib-metadata==5.2.0
importlib-resources==5.10.2
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jax==0.3.6
jaxlib==0.3.5
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mmcls==0.25.0
mmcv-full==1.7.0
model-index==0.1.11
msgpack==1.0.4
mysql-connector-python==8.0.33
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.0
nvidia-ml-py==11.495.46
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openmim==0.3.3
opt-einsum==3.3.0
optax==0.1.4
optuna==3.1.1
optuna-dashboard==0.9.2
ordered-set==4.1.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.3
prettytable==3.5.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyMySQL==1.0.3
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==11.2.0
rsa==4.9
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.13
stack-data==0.6.2
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorstore==0.1.28
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.13.2
toolz==0.12.0
torch==1.8.1+cu101
torchaudio==0.8.1
torchvision==0.9.1+cu101
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
transformers @ git+https://github.com/huggingface/transformers@539e2281cd97c35ef4122757f26c88f44115fa94
transforms==0.1
ttach==0.0.3
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
virtualenv==20.23.0
wandb==0.13.7
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
yapf==0.32.0
zipp==3.11.0
